+++ src/os.h
@@ -11,7 +11,8 @@
 typedef uint32_t os_stack_t;
 
 os_error_t os_init(void);
-os_error_t os_task_init(void (*handler)(void), os_stack_t *p_stack, uint32_t stack_size);
+os_error_t os_task_init(void (*handler)(void), os_stack_t *p_stack,
+			size_t stack_size);
 os_error_t os_start(uint32_t systick_ticks);
 
 #endif
